+++ internal/model/encoding_test.go
@@ -12,13 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package translator
+package model
 
 import (
-	"fmt"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
-// NewErrIncompatibleType returns errIncompatibleType instance
-func NewErrIncompatibleType(expected, given interface{}) error {
-	return fmt.Errorf("expected model type %T but given %T", expected, given)
+func TestNewErrUnavailableEncoding(t *testing.T) {
+	err := NewErrUnavailableEncoding("unknown")
+	assert.IsType(t, &ErrUnavailableEncoding{}, err)
+	assert.EqualError(t, err, `unsupported encoding "unknown"`)
+}
+
+func TestEncoding_String(t *testing.T) {
+	assert.Equal(t, "protobuf", Protobuf.String())
+	assert.Equal(t, "json", JSON.String())
+	assert.Equal(t, "thrift", Thrift.String())
 }