+++ internal/model/unmarshal_test.go
@@ -23,19 +23,19 @@ import (
 	"go.opentelemetry.io/collector/consumer/pdata"
 )
 
-func TestTracesDecoder_SerializeError(t *testing.T) {
-	translate := &mockTranslator{}
-	serialize := &mockSerializer{}
+func TestTracesUnmarshal_EncodingError(t *testing.T) {
+	translator := &mockTranslator{}
+	encoder := &mockEncoder{}
 
 	d := &TracesUnmarshaler{
-		translate: translate,
-		serialize: serialize,
+		translator: translator,
+		encoder:    encoder,
 	}
 
 	expectedBytes := []byte{1, 2, 3}
 	expectedModel := struct{}{}
 
-	serialize.On("UnmarshalTraces", expectedBytes).Return(expectedModel, errors.New("decode failed"))
+	encoder.On("DecodeTraces", expectedBytes).Return(expectedModel, errors.New("decode failed"))
 
 	_, err := d.Unmarshal(expectedBytes)
 
@@ -43,20 +43,20 @@ func TestTracesDecoder_SerializeError(t *testing.T) {
 	assert.EqualError(t, err, "unmarshal failed: decode failed")
 }
 
-func TestTracesDecoder_TranslationError(t *testing.T) {
-	translate := &mockTranslator{}
-	serialize := &mockSerializer{}
+func TestTracesUnmarshal_TranslationError(t *testing.T) {
+	translator := &mockTranslator{}
+	encoder := &mockEncoder{}
 
 	d := &TracesUnmarshaler{
-		translate: translate,
-		serialize: serialize,
+		translator: translator,
+		encoder:    encoder,
 	}
 
 	expectedBytes := []byte{1, 2, 3}
 	expectedModel := struct{}{}
 
-	serialize.On("UnmarshalTraces", expectedBytes).Return(expectedModel, nil)
-	translate.On("DecodeTraces", expectedModel).Return(pdata.NewTraces(), errors.New("translation failed"))
+	encoder.On("DecodeTraces", expectedBytes).Return(expectedModel, nil)
+	translator.On("ToTraces", expectedModel).Return(pdata.NewTraces(), errors.New("translation failed"))
 
 	_, err := d.Unmarshal(expectedBytes)
 
@@ -64,21 +64,21 @@ func TestTracesDecoder_TranslationError(t *testing.T) {
 	assert.EqualError(t, err, "converting model to pdata failed: translation failed")
 }
 
-func TestTracesDecoder_Decode(t *testing.T) {
-	translate := &mockTranslator{}
-	serialize := &mockSerializer{}
+func TestTracesUnmarshal_Decode(t *testing.T) {
+	translator := &mockTranslator{}
+	encoder := &mockEncoder{}
 
 	d := &TracesUnmarshaler{
-		translate: translate,
-		serialize: serialize,
+		translator: translator,
+		encoder:    encoder,
 	}
 
 	expectedTraces := pdata.NewTraces()
 	expectedBytes := []byte{1, 2, 3}
 	expectedModel := struct{}{}
 
-	serialize.On("UnmarshalTraces", expectedBytes).Return(expectedModel, nil)
-	translate.On("DecodeTraces", expectedModel).Return(expectedTraces, nil)
+	encoder.On("DecodeTraces", expectedBytes).Return(expectedModel, nil)
+	translator.On("ToTraces", expectedModel).Return(expectedTraces, nil)
 
 	actualTraces, err := d.Unmarshal(expectedBytes)